# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Favin Wallet",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/all-wallets": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Get list of all Wallets",
                "description": "Returns list of Wallets",
                "operationId": "getAllWalletsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\WalletCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/user-wallets": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Get list of all User Wallets",
                "description": "Returns list of User Wallets",
                "operationId": "getAllUserWalletsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\WalletCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "No Wallets Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/wallets": {
            "post": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Store new Wallet",
                "description": "Returns Wallet data",
                "operationId": "storeWallets",
                "requestBody": {
                    "description": "Pass Wallet credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "new wallet"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "new wallet"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\WalletResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}